refactor(Query): move `Query` into its own file
